# Vectorise expand8 in ForUtil using JDK Vector API